% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R, R/filter.R
\name{classify_tile}
\alias{classify_tile}
\alias{filter_annual_samples}
\title{Classify a Data Cube tile}
\usage{
classify_tile(cube, model, ..., smooth_type = "bayes")

filter_annual_samples(samples_ts)
}
\arguments{
\item{...}{...}
}
\value{
todo

todo
}
\description{
description

description
}
